pyproject.toml
[tool.poetry]
name = "procgenart"
version = "0.1.0"
description = ""
authors = ["Chris Dirkis <git@chrisdirk.is>"]

[tool.poetry.dependencies]
python = "^3.8"
Pillow = "^8.4.0"
tweepy = "^4.1.0"
python-dotenv = "^0.19.1"
xxhash = "^2.0.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
